Allow AlgoConfiguration to configure the Algo
Pass the Algo to the AlgoConfiguration and have the configuration correctly configure the Algo.

All the steps are executed in the provided order.
REVERSE => algo.reverse()
SET_UP => algo.setUp()
SET_ALGO_PARAM => algo.setParam(param, value) for each of the AlgoParamValues configured against the AlgoConfiguration
PERFORM_CALC => algo.performCalc()
SUBMIT_TO_MARKET => ago.submitToMarket()

Added test class for AlgoConfiguration to ensure that it is being created correctly from the entity, and then correctly applying the configuration to the Algo